Rename Serializer.hpp to NetworkHelper.hpp
Fix a tests error and  some typos.
